Define a Set of Multivariate Multidimensional Functional Basis

The `mvbasismfd` a set of multivariate multidimensional basis functions. This class utilizes basis objects `basismfd`.

Constructor for `mvbasismfd` objects (same as `Mvbasismfd`)

Usage

Mvbasismfd(basis)

Mvbasismfd(basis)

# S3 method for class 'mvbasismfd'
mvbasismfd_obj[i = "index"]

Arguments

basis

A list of basisfd objects

mvbasismfd_obj

An 'mvmfd' object

i

An index or indices specifying the subsets to extract for the first dimension

Value

An `mvbasismfd` object containing the specified subsets

See also

Active bindings

nvar

number of variables

basis

A list of `mvbasisfd` objects

dimSupp

A sequence of positive integers specifying support domain of the `mvbasismfd` object.

nbasis

A list of integers specifying the number of basis functions

supp

A list of matrices specifying the support of basis functions

gram

The Gram matrix.

Methods

Public methods

mvbasismfd$new()

mvbasismfd$eval()

mvbasismfd$clone()

Method new()

Constructor for `mvbasismfd` objects (same as Mvbasismfd(...) )

Usage

mvbasismfd$new(basis)

Arguments

basis

A list of `basismfd` objects

Method eval()

Evaluate the `mvbasismfd` object at given argument values

Usage

mvbasismfd$eval(evalarg)

Arguments

evalarg

A list of numeric vectors of argument values at which the `mvbasismfd` is to be evaluated

Returns

A list of evaluated values

Method clone()

The objects of this class are cloneable with this method.

Usage

mvbasismfd$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.