Scale a `mvmfd` Object

This function scales a `mvmfd` object by scaling each functional variable independently using either a calculated or user-provided weight. It returns a new scaled `mvmfd` object.

Usage

scale_mvmfd(
  mvmfd_obj,
  mvmfd_eval_length = rep(100, mvmfd_obj$nvar),
  weight = NULL
)

Arguments

mvmfd_obj

An object of class `mvmfd`.

mvmfd_eval_length

A numeric vector specifying the number of evaluation points for each functional variable.

weight

An optional numeric vector of scaling factors for each functional variable. If NULL, scaling factors are calculated automatically.

Value

A scaled mvmfd object.

Examples

# Example usage:
# Assuming `mvmfd_obj` is a valid mvmfd object:
# scaled_mvmfd <- scale_mvmfd(mvmfd_obj, mvmfd_eval_length = c(100, 200))
# scaled_mvmfd <- scale_mvmfd(mvmfd_obj, mvmfd_eval_length = c(100, 200), weight = c(0.5, 0.8))